Renamed BatteryDrainThread to BatteryUpdateThread that is more generic Moved peeking battery information there One loop for each supported platform: on windows use Kernel32 API, and on others use the draining code. Prepared linux code (fetching /sys/class/power_supply/BAT0 information from folder for new windows systems)
Removed log file that I uploaded as an error
