+++ .gitignore
@@ -35,3 +35,5 @@ ms0
 .gradle
 step-trace.txt
 .settings
+capture.bin
+hs*.log
